Another redefining of the vehicle in frint check
The previous redefining did't fully address the problem caused by the
IDM-function which resulted in vehicles driving "over" vehicle in
front. The new "vehicle in front"-check uses the previously existing
private function checkCurrentBraking(int). In case this function
returns the integer 1 a vehicle in front is found. This new boolean
expressions seems to fix teh overrunning bug the IDM-function was
previously dealing with.