Trafficmodel: Replacing a variable in IDM function
The IDM function needs a variable called  desired velocity in the
papers. The porevious implementation considered this desired velocity
as the maximum speed allowed on the actual street. To make the IDM
function work more realistic the desired velocity is now the minimum
of the maximum speed allowed for the actual street and the maximum
speed of the actual car. Prior to this change the maximum speed of
the car was modeled by a hard limit check. Now the IDM function itself
regulate this hard limit and simulate real world situation far better.