Trafficmodel: Changed the GPS->MapMatric function
The function translateGPSToMapMetric(int[],double,double) is used to
compute corresponding coordinates in the map metric of Vanet-Sim from
original latitude and longitudes values. Therefore a very simple
mathmatic function was used. This implementation however doesn't paid
attention to the fact, that maps may vary in their height and width
dimensions without altering the spatial node relation in the map.
The new implementation now search for the mimimum/maximum nodes to
map map the max/min longitude/latitudes to these found nodes. With
this the free spatial space at the boundrys of the map can be ignored
while computing the map metric values from the GPS values.