Trafficmodel: Started the impl. of VehicleMaster
For inserting GPS-created vehicles on the right time, a class for
managing the vehciles, which will be inserted. This class is called
GPSVehicleMaster. This class inserts the vehicles into the map, when
the time is right. This class uses a custom Queueimplementation, which
stores the vehicles sorted by their starttime. By adding a vehicle it
is automatically inserted on the right place within the queue. The
only way to remove an vehicle is by removing the first vehicle in the
queue.